+++ src/avancedb/json_stream.h
@@ -11,7 +11,7 @@ public:
     void Append(const char* name, const std::string& value);
     void Append(const char* name, const char* value);
     void Append(const char* name, int value);
-    void Append(const char* name, unsigned value);
+    void Append(const char* name, unsigned long value);
     void Append(const char* name, uint64_t value);
     void Append(const char* name, int64_t value);
     void Append(const char* name, double value);
